Add an integration test to exercise Puck3D deinit
