+++ Tests/MapboxMapsTests/Location/Pucks/Puck3DIntegrationTests.swift
@@ -1,0 +1,26 @@
+import XCTest
+import MapboxMaps
+
+final class Puck3DIntegrationTests: XCTestCase {
+
+    var resourceOptions: ResourceOptions!
+
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+        resourceOptions = try ResourceOptions(accessToken: mapboxAccessToken())
+    }
+
+    func testPuck3DDeinitializationDoesNotCrash() {
+        autoreleasepool {
+            let mapView = MapView(with: .zero, resourceOptions: resourceOptions)
+            mapView.update { (options) in
+                options.location.puckType = .puck3D(
+                    Puck3DConfiguration(
+                        model: Model()))
+            }
+        }
+        // there is no assertion here because this test is
+        // merely ensuring that Puck3D does not crash when
+        // it gets deinited
+    }
+}
